Added Shop Scene
Added shop scene, features include:

Shop building with shelves and buttons on each shelf for purchasing items.

Placeholder potion items which bob up and down.

Dictionary for all items in the game (can be expanded upon).

When an item is clicked (purchased), the ID attached to the item is added to the players inventory.

When I is held down, the player can view their inventory, including # of each item in inventory.

To-do: Add restrictions on purchase amount without destroying game object.